* https://github.com/osm0sis/AnyKernel2:
  README: cosmetic changes to match XDA thread
  AK2: underp opposite slot case
  README: update for multi-slot zip instructions
  AK2: add support for multi-slot zips - extend `reset_ak` with `reset_ak keep` which will keep files to be available to patch the other slot - add optional slot_select=inactive shell variable to force inactive slot during slot detection
  AK2: add U-Boot RAMDisk uImage support
  README: update for Intel OSIP mboot, customdd, touchups
  AK2: add $OS$ header Android image format support - add built-in support for mboot for Intel OSIP devices (see: https://github.com/osm0sis/mboot)
  AK2: fix dump failures for unpackelf and dumpimage (with ramdisk)

+++ tools/ak2-core.sh
@@ -22,14 +22,21 @@ contains() { test "${1#*$2}" != "$1" && return 0 || return 1; }
 # file_getprop <file> <property>
 file_getprop() { grep "^$2=" "$1" | cut -d= -f2-; }
 
-# reset anykernel directory
+# reset_ak [keep]
 reset_ak() {
   local i;
-  rm -rf $(dirname /tmp/anykernel/*-files/current)/ramdisk;
-  for i in $ramdisk $split_img /tmp/anykernel/rdtmp /tmp/anykernel/boot.img /tmp/anykernel/*-new*; do
-    cp -af $i $(dirname /tmp/anykernel/*-files/current);
-  done;
-  rm -rf $ramdisk $split_img $patch /tmp/anykernel/rdtmp /tmp/anykernel/boot.img /tmp/anykernel/*-new* /tmp/anykernel/*-files/current;
+  if [ ! "$1" == "keep" ]; then
+    rm -rf $(dirname /tmp/anykernel/*-files/current)/ramdisk;
+    for i in $ramdisk $split_img /tmp/anykernel/rdtmp /tmp/anykernel/boot.img /tmp/anykernel/*-new*; do
+      cp -af $i $(dirname /tmp/anykernel/*-files/current);
+    done;
+  fi;
+  rm -rf $ramdisk $split_img /tmp/anykernel/boot.img /tmp/anykernel/*-new* /tmp/anykernel/*-files/current;
+  if [ "$1" == "keep" ]; then
+    mv -f /tmp/anykernel/rdtmp $ramdisk;
+  else
+    rm -rf $patch /tmp/anykernel/rdtmp;
+  fi;
   . /tmp/anykernel/tools/ak2-core.sh $FD;
 }
 
@@ -67,7 +74,14 @@ split_boot() {
       cp -f $split_img/elftool_out/header $split_img/boot.img-header;
     fi;
     $bin/unpackelf -i /tmp/anykernel/boot.img -o $split_img;
+    test $? != 0 && dumpfail=1;
     mv -f $split_img/boot.img-ramdisk.cpio.gz $split_img/boot.img-ramdisk.gz;
+  elif [ -f "$bin/mboot" ]; then
+    $bin/mboot -u -f /tmp/anykernel/boot.img -d $split_img;
+    test $? != 0 && dumpfail=1;
+    mv -f $split_img/cmdline.txt $split_img/boot.img-cmdline;
+    mv -f $split_img/kernel $split_img/boot.img-zImage;
+    mv -f $split_img/ramdisk.cpio.gz $split_img/boot.img-ramdisk.gz;
   elif [ -f "$bin/dumpimage" ]; then
     uimgsize=$(($(printf '%d\n' 0x$(hexdump -n 4 -s 12 -e '16/1 "%02x""\n"' /tmp/anykernel/boot.img)) + 64));
     if [ "$(wc -c < /tmp/anykernel/boot.img)" != "$uimgsize" ]; then
@@ -85,9 +99,10 @@ split_boot() {
     grep "Point:" $split_img/boot.img-header | cut -c15- > $split_img/boot.img-ep;
     $bin/dumpimage -p 0 -o $split_img/boot.img-zImage /tmp/anykernel/boot.img;
     test $? != 0 && dumpfail=1;
-    if [ "$(cat $split_img/boot.img-type)" == "Multi" ]; then
-      $bin/dumpimage -p 1 -o $split_img/boot.img-ramdisk.gz /tmp/anykernel/boot.img;
-    fi;
+    case $(cat $split_img/boot.img-type) in
+      Multi) $bin/dumpimage -p 1 -o $split_img/boot.img-ramdisk.gz /tmp/anykernel/boot.img;;
+      RAMDisk) mv -f $split_img/boot.img-zImage $split_img/boot.img-ramdisk.gz;;
+    esac;
     test $? != 0 && dumpfail=1;
   elif [ -f "$bin/rkcrc" ]; then
     dd bs=4096 skip=8 iflag=skip_bytes conv=notrunc if=/tmp/anykernel/boot.img of=$split_img/boot.img-ramdisk.gz;
@@ -205,7 +220,7 @@ flash_dtbo() {
   fi;
 }
 flash_boot() {
-  local name arch os type comp addr ep cmdline cmd board base pagesize kerneloff ramdiskoff tagsoff dtboff osver oslvl hdrver second secondoff recoverydtbo hash unknown i kernel rd dtb dt rpm pk8 cert avbtype dtbo dtbo_block;
+  local name arch os type comp addr ep cmdline cmd board base pagesize kerneloff ramdiskoff tagsoff dtboff osver oslvl hdrver second secondoff recoverydtbo hash unknown i kernel rd dtb dt rpm part0 part1 pk8 cert avbtype dtbo dtbo_block;
   cd $split_img;
   if [ -f "$bin/mkimage" ]; then
     name=`cat *-name`;
@@ -309,10 +324,19 @@ flash_boot() {
     esac;
   fi;
   if [ -f "$bin/mkimage" ]; then
-    test "$type" == "Multi" && uramdisk=":$rd";
-    $bin/mkimage -A $arch -O $os -T $type -C $comp -a $addr -e $ep -n "$name" -d $kernel$uramdisk boot-new.img;
+    part0=$kernel;
+    case $type in
+      Multi) part1=":$rd";;
+      RAMDisk) part0=$rd;;
+    esac;
+    $bin/mkimage -A $arch -O $os -T $type -C $comp -a $addr -e $ep -n "$name" -d $part0$part1 boot-new.img;
   elif [ -f "$bin/elftool" ]; then
     $bin/elftool pack -o boot-new.img header=$split_img/boot.img-header $kernel $rd,ramdisk $rpm $cmd;
+  elif [ -f "$bin/mboot" ]; then
+    cp -f $split_img/boot.img-cmdline $split_img/cmdline.txt;
+    cp -f $kernel $split_img/kernel;
+    cp -f $rd $split_img/ramdisk.cpio.gz;
+    $bin/mboot -d $split_img -f boot-new.img;
   elif [ -f "$bin/rkcrc" ]; then
     $bin/rkcrc -k $rd boot-new.img;
   elif [ -f "$bin/pxa-mkbootimg" ]; then
@@ -591,6 +615,16 @@ case $is_slot_device in
       slot=$(getprop ro.boot.slot 2>/dev/null);
       test ! "$slot" && slot=$(grep -o 'androidboot.slot=.*$' /proc/cmdline | cut -d\  -f1 | cut -d= -f2);
       test "$slot" && slot=_$slot;
+    fi;
+    if [ "$slot" ]; then
+      case $slot_select in
+        inactive)
+          case $slot in
+            _a) slot=_b;;
+            _b) slot=_a;;
+          esac;
+        ;;
+      esac;
     fi;
     if [ ! "$slot" -a "$is_slot_device" == 1 ]; then
       ui_print " "; ui_print "Unable to determine active boot slot. Aborting..."; exit 1;
